requirements.txt
streamlit
google-generativeai
speechrecognition
pyaudio